file splitting and stats files
