+++ cmd/scraper/scraper.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"bactic/internal/database"
-	"bactic/internal/scrapers/athnet"
 	"bactic/internal/scrapers/tfrrs"
+	"context"
 	"flag"
 	"log"
 	"os"
@@ -11,6 +11,7 @@ import (
 	"strings"
 	"sync"
 	"syscall"
+	"time"
 
 	_ "github.com/lib/pq"
 )
@@ -18,16 +19,23 @@ import (
 func main() {
 	var (
 		scrapersList string
+		scrapeInt    time.Duration
 		dbURL        string
 		found        bool
 		verbosity    int
 	)
-	validScrapers := map[string](func(*database.BacticDB, chan os.Signal, *sync.WaitGroup)){"tfrrs": tfrrs.NewTFRRSCollector, "athnet": athnet.NewAthnetCollector}
+	validScrapers := map[string](func(*database.BacticDB, context.Context, *sync.WaitGroup, time.Duration)){
+		"tfrrs": tfrrs.NewTFRRSScraper,
+		// "athnet": athnet.NewAthnetCollector,
+	}
 
 	flag.StringVar(&scrapersList, "scrapers", "tfrrs", "Comma-separated list of scrapers to run concurrently. Any of \"tfrrs\" and \"athnet\"")
 	flag.StringVar(&dbURL, "db", "", "Fully-qualified postgres url. Overrides the environment variable defined in DB_URL")
 	flag.IntVar(&verbosity, "verbosity", 1, "verbosity level (1, 2, 3)")
+	flag.DurationVar(&scrapeInt, "duration", time.Hour*24, "Interval between scrapes")
 	flag.Parse()
+
+	log.SetPrefix("Scraper main")
 
 	if len(dbURL) == 0 {
 		dbURL, found = os.LookupEnv("DB_URL")
@@ -40,7 +48,7 @@ func main() {
 	db.SetupSchema()
 	defer db.Close()
 
-	var scraperSet []func(*database.BacticDB, chan os.Signal, *sync.WaitGroup)
+	var scraperSet []func(*database.BacticDB, context.Context, *sync.WaitGroup, time.Duration)
 
 	for _, s := range strings.Split(scrapersList, ",") {
 		scraper, found := validScrapers[s]
@@ -51,17 +59,19 @@ func main() {
 	}
 
 	interrupt := make(chan os.Signal, 1)
+	ctx, cancel := context.WithCancel(context.Background())
 	signal.Notify(interrupt, syscall.SIGINT)
 	var wg sync.WaitGroup
 
 	for _, startScraper := range scraperSet {
 		wg.Add(1)
-		go startScraper(db, interrupt, &wg)
+		go startScraper(db, ctx, &wg, scrapeInt)
 	}
 
 	go func() {
 		<-interrupt
 		log.Println("Received interrupt signal, shutting down existing scrapers...")
+		cancel()
 	}()
 
 	wg.Wait()
