+++ internal/scrapers/athnet/athnet_scraper.go
@@ -5,8 +5,9 @@ import (
 	"log"
 	"os"
 	"sync"
+	"time"
 )
 
-func NewAthnetCollector(db *database.BacticDB, sig chan os.Signal, wg *sync.WaitGroup) {
+func NewAthnetCollector(db *database.BacticDB, sig chan os.Signal, wg *sync.WaitGroup, scrapeLoop time.Duration) {
 	log.Fatal("Athnet collector not yet implemented")
 }
