+++ internal/stats/computer.go
@@ -1,0 +1,4 @@
+package stats
+
+type StatsComputer struct {
+}
