+++ internal/stats/historgram.go
@@ -1,0 +1,10 @@
+package stats
+
+type Histogram struct {
+}
+
+func (h *Histogram) Compute() {
+}
+
+func (h *Histogram) Query() {
+}
