+++ internal/stats/stat.go
@@ -1,0 +1,7 @@
+package stats
+
+// An interface for computable and query-capable stats
+type Stat interface {
+	Compute(string) error
+	Query(string) interface{}
+}
